Add ctest unit tests (#15)
* latest submod

* checkpoint before adding tests.

* testing working

* add more tests

* version opt

* fix lint warnings

* windows fixes

* fixes for Ubuntu

* submod

* Move ignoring C++ warnings explicitly to when including ezgz.hpp

* add testing to build step

* updates to try to find tests

* update build workflow for tests

* update build workflow

* add some instrumentation for testing

* allow vector to checkStderr and checkStdout

* wip (tests)

* testing instrumentation working

* refactor test args for easier setup

* win changes

* massage options check

* add data directory for tests

* first data tests

* add comparison to reference files

* winfix

* export tests code complete

* updated test files

* wip xfer to windows

* winfix

* another winfix

* another winfix

* tests completed except for logging

* check if it works on GitHub

* log tests added

* submod

+++ src/utils/score_encoder.h
@@ -70,7 +70,7 @@ public:
             // this algorithm is BSD rand()!
             state = state * 0x41c64e6d + 0x3039;
             uint16_t upper = state >> 16;
-            uint8_t c = upper + upper / 255;
+            uint8_t c = uint8_t(upper + upper / 255);
             buffer[i] ^= c;
         }
     }
